src/requirements.txt
pysha3

pandas~=1.1.3
mysql-connector-python~=8.0.22
flask~=1.1.2
SQLAlchemy~=1.3.20
MarkupSafe~=1.1.1